extract the agg package
